move page tests into isolated directory
Next tries to build any .tsx-file in the pages directory, but we don't
want our tests to be considered pages.
